# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pulse API
  description: Self-hosted uptime monitoring platform API.
  version: 1.0.0
  contact:
    name: Pulse
servers:
  - url: /api/v1
    description: API v1
security:
  - bearerAuth: []
paths:
...
  /monitors/{id}:
    get:
      operationId: getMonitor
      summary: Get a monitor by ID
      tags: [Monitors]
      parameters:
        - $ref: "#/components/parameters/MonitorID"
      responses:
        "200":
          description: Monitor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: putMonitor
      summary: Create or update a monitor (idempotent)
      tags: [Monitors]
      parameters:
        - $ref: "#/components/parameters/MonitorID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutMonitorRequest"
      responses:
        "200":
          description: Monitor created or updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteMonitor
      summary: Delete a monitor
      tags: [Monitors]
      parameters:
        - $ref: "#/components/parameters/MonitorID"
      responses:
        "204":
          description: Monitor deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
  /monitors/{id}/history:
    get:
      operationId: getMonitorHistory
      summary: Get check history for a monitor
      tags: [Monitors]
      parameters:
        - $ref: "#/components/parameters/MonitorID"
        - name: from
          in: query
          description: Start of time range (RFC 3339). Defaults to 24h ago.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: End of time range (RFC 3339). Defaults to now.
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Check history points
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /monitors/{id}/credentials:
    post:
      operationId: createMonitorCredential
      summary: Create a credential for a monitor
      description: >
        Creates a new authentication credential for the specified monitor.
        The credential value is encrypted at rest and never returned in subsequent API responses.
      tags: [Credentials]
      parameters:
        - $ref: "#/components/parameters/MonitorID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMonitorCredentialRequest"
      responses:
        "201":
          description: Credential created (secret values not returned)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorCredential"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listMonitorCredentials
      summary: List credentials for a monitor
      description: >
        Returns metadata for all credentials associated with the specified monitor.
        Secret values are never included in the response.
      tags: [Credentials]
      parameters:
        - $ref: "#/components/parameters/MonitorID"
      responses:
        "200":
          description: List of credential metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorCredentialListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /monitors/{id}/credentials/{credentialId}:
    put:
      operationId: updateMonitorCredential
      summary: Update a monitor credential
      description: >
        Updates an existing credential. Only the provided fields are updated.
        The new secret value is encrypted at rest and never returned in the response.
      tags: [Credentials]
      parameters:
        - $ref: "#/components/parameters/MonitorID"
        - $ref: "#/components/parameters/CredentialID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMonitorCredentialRequest"
      responses:
        "200":
          description: Credential updated (secret values not returned)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorCredential"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMonitorCredential
      summary: Delete a monitor credential
      description: >
        Permanently removes a credential from the monitor. The monitor will stop
        including this credential's authentication in subsequent health-check requests.
      tags: [Credentials]
      parameters:
        - $ref: "#/components/parameters/MonitorID"
        - $ref: "#/components/parameters/CredentialID"
      responses:
        "204":
          description: Credential deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or API Token
  parameters:
...
    MonitorID:
      name: id
      in: path
      required: true
      description: Monitor UUID
      schema:
        type: string
        format: uuid
    CredentialID:
      name: credentialId
      in: path
      required: true
      description: Credential UUID
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    Monitor:
      type: object
      required:
        - id
        - name
        - type
        - target
        - interval_seconds
        - timeout_seconds
        - status
        - state
        - settings
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [http, tcp, udp, websocket]
        target:
          type: string
        interval_seconds:
          type: integer
        timeout_seconds:
          type: integer
        status:
          type: string
          enum: [active, paused]
        state:
          type: string
          enum: [up, down, unknown]
        last_checked_at:
          type: string
          format: date-time
          nullable: true
        next_check_at:
          type: string
          format: date-time
          nullable: true
        settings:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PutMonitorRequest:
      type: object
      required: [name, type, target]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [http, tcp, udp, websocket]
        target:
          type: string
        interval_seconds:
          type: integer
          default: 60
        timeout_seconds:
          type: integer
          default: 10
        status:
          type: string
          enum: [active, paused]
          default: active
        settings:
          type: object
...
    HistoryPoint:
      type: object
      required: [state, checked_at]
      properties:
        state:
          type: string
          enum: [up, down]
        latency_ms:
          type: integer
          nullable: true
        status_code:
          type: integer
          nullable: true
        error:
          type: string
          nullable: true
        ssl_days_remaining:
          type: integer
          nullable: true
          description: Days until SSL certificate expires (HTTPS monitors only)
        checked_at:
          type: string
          format: date-time
    HistoryResponse:
      type: object
      required: [monitor_id, from, to, points]
      properties:
        monitor_id:
          type: string
          format: uuid
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        points:
          type: array
          items:
            $ref: "#/components/schemas/HistoryPoint"
...
    MonitorCredential:
      type: object
      description: >
        Credential metadata for a monitor. Secret values (token, password, header_value,
        encrypted_value) are never included in API responses.
      required: [id, auth_type, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        auth_type:
          type: string
          enum: [bearer, basic, header]
          description: The type of authentication this credential provides
        name:
          type: string
          description: Human-readable label for the credential
        header_name:
          type: string
          nullable: true
          description: Custom header name (only present when auth_type is header)
        username:
          type: string
          nullable: true
          description: Username for basic auth (only present when auth_type is basic)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateMonitorCredentialRequest:
      type: object
      required: [auth_type, name]
      properties:
        auth_type:
          type: string
          enum: [bearer, basic, header]
          description: The type of authentication this credential provides
        name:
          type: string
          description: Human-readable label for the credential
        token:
          type: string
          description: Bearer token value (required when auth_type is bearer)
        username:
          type: string
          description: Username for basic auth (required when auth_type is basic)
        password:
          type: string
          description: Password for basic auth (required when auth_type is basic)
        header_name:
          type: string
          description: Custom header name (required when auth_type is header)
        header_value:
          type: string
          description: Custom header value (required when auth_type is header)
    UpdateMonitorCredentialRequest:
      type: object
      properties:
        name:
          type: string
          description: Updated human-readable label for the credential
        token:
          type: string
          description: New bearer token value (for auth_type bearer)
        username:
          type: string
          description: New username for basic auth (for auth_type basic)
        password:
          type: string
          description: New password for basic auth (for auth_type basic)
        header_name:
          type: string
          description: New custom header name (for auth_type header)
        header_value:
          type: string
          description: New custom header value (for auth_type header)
    MonitorCredentialListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MonitorCredential"
...