# backend/internal/store/postgres/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"encoding/json"
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type ApiToken struct {
	ID         uuid.UUID          `db:"id" json:"id"`
	UserID     uuid.UUID          `db:"user_id" json:"user_id"`
	Name       string             `db:"name" json:"name"`
	TokenHash  string             `db:"token_hash" json:"token_hash"`
	LastUsedAt pgtype.Timestamptz `db:"last_used_at" json:"last_used_at"`
	ExpiresAt  pgtype.Timestamptz `db:"expires_at" json:"expires_at"`
	RevokedAt  pgtype.Timestamptz `db:"revoked_at" json:"revoked_at"`
	CreatedAt  time.Time          `db:"created_at" json:"created_at"`
	Prefix     string             `db:"prefix" json:"prefix"`
}

type CheckResult struct {
	ID               uuid.UUID `db:"id" json:"id"`
	MonitorID        uuid.UUID `db:"monitor_id" json:"monitor_id"`
	CheckedAt        time.Time `db:"checked_at" json:"checked_at"`
	State            string    `db:"state" json:"state"`
	LatencyMs        *int32    `db:"latency_ms" json:"latency_ms"`
	StatusCode       *int32    `db:"status_code" json:"status_code"`
	Error            *string   `db:"error" json:"error"`
	SslDaysRemaining *int32    `db:"ssl_days_remaining" json:"ssl_days_remaining"`
	CreatedAt        time.Time `db:"created_at" json:"created_at"`
}

type Incident struct {
	ID         uuid.UUID          `db:"id" json:"id"`
	MonitorID  uuid.UUID          `db:"monitor_id" json:"monitor_id"`
	StartedAt  time.Time          `db:"started_at" json:"started_at"`
	ResolvedAt pgtype.Timestamptz `db:"resolved_at" json:"resolved_at"`
	Cause      *string            `db:"cause" json:"cause"`
	CreatedAt  time.Time          `db:"created_at" json:"created_at"`
}

type Monitor struct {
	ID              uuid.UUID          `db:"id" json:"id"`
	Name            string             `db:"name" json:"name"`
	Type            string             `db:"type" json:"type"`
	Target          string             `db:"target" json:"target"`
	IntervalSeconds int32              `db:"interval_seconds" json:"interval_seconds"`
	TimeoutSeconds  int32              `db:"timeout_seconds" json:"timeout_seconds"`
	Status          string             `db:"status" json:"status"`
	State           string             `db:"state" json:"state"`
	LastCheckedAt   pgtype.Timestamptz `db:"last_checked_at" json:"last_checked_at"`
	NextCheckAt     pgtype.Timestamptz `db:"next_check_at" json:"next_check_at"`
	Settings        json.RawMessage    `db:"settings" json:"settings"`
	CreatedAt       time.Time          `db:"created_at" json:"created_at"`
	UpdatedAt       time.Time          `db:"updated_at" json:"updated_at"`
}

type MonitorCredential struct {
	ID             uuid.UUID `db:"id" json:"id"`
	MonitorID      uuid.UUID `db:"monitor_id" json:"monitor_id"`
	AuthType       string    `db:"auth_type" json:"auth_type"`
	Name           string    `db:"name" json:"name"`
	EncryptedValue string    `db:"encrypted_value" json:"encrypted_value"`
	CreatedAt      time.Time `db:"created_at" json:"created_at"`
	UpdatedAt      time.Time `db:"updated_at" json:"updated_at"`
}

type Secret struct {
	ID             uuid.UUID `db:"id" json:"id"`
	Name           string    `db:"name" json:"name"`
	EncryptedValue string    `db:"encrypted_value" json:"encrypted_value"`
	CreatedAt      time.Time `db:"created_at" json:"created_at"`
	UpdatedAt      time.Time `db:"updated_at" json:"updated_at"`
}

type User struct {
	ID           uuid.UUID `db:"id" json:"id"`
	Email        string    `db:"email" json:"email"`
	PasswordHash string    `db:"password_hash" json:"password_hash"`
	CreatedAt    time.Time `db:"created_at" json:"created_at"`
	UpdatedAt    time.Time `db:"updated_at" json:"updated_at"`
}